Refactor how we deal with metrics.
EvoLogger is primarily focussed on providing information solely for the
Honeywell Evohome platform.  However, we want to extend the capabilities
of this further, to allow a more broad-spectrum of metrics gathering.

For Example:

Extend capabilities of DarkSky API data returned and stored.
Allow further plugins to be utilised (DCC consumption metrics?)

For this, we are removing the focus on the 'Temperature' metric in favour
of a multi-purpose approach.

The series definition will also include which Plugin has created the metric,
as well as sanitise the descriptor (previously zone).

The focus for this is to fine tune it more to how I would personally like to
retrieve and store metrics within Influx, for graphical representaion in
Grafana.