Add Cron-style Polling Interval support.
Previously, Polling Intervals were set as seconds, and by default (in config.ini) was
set to 10 minutes (600s).

This change will allow the use of both cron-style timers (default) or the ability to specify
arbituary values (seconds).  This allows for more control on when a resource can be polled,
so could be at set times within the hour/day, instead of blindly using sleep intervals.

If an interger value is specified as an interval, and it is less than 60, it will produce a
warning and then default to 60s, so to avoide overloading any 3rd party endpoints or other
services.

Going forward, if running multiple input plugins, it may be beneficial to allow polling to
move away from the main evologger process, and be configurable per-plugin.  This will be the
next change, where plugins will support polling intervals. The default value of which will be
every minute.